no internet case + comments
