+++ app/src/main/java/com/tmvlg/factorcapgame/FactOrCapApplication.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import com.tmvlg.factorcapgame.data.FactOrCapDatabase
 import com.tmvlg.factorcapgame.data.preferences.PreferenceProvider
 import com.tmvlg.factorcapgame.data.repository.fact.FactRepository
-import com.tmvlg.factorcapgame.data.repository.game.GameRepositoryImpl
+import com.tmvlg.factorcapgame.data.repository.game.GameRepository
 import com.tmvlg.factorcapgame.data.repository.user.UserRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
@@ -13,7 +13,7 @@ class FactOrCapApplication : Application() {
     val applicationScope = CoroutineScope(SupervisorJob())
 
     val database by lazy { FactOrCapDatabase.getDatabase(this, applicationScope) }
-    val gameRepository by lazy { GameRepositoryImpl(database.gameDao()) }
+    val gameRepository by lazy { GameRepository(database.gameDao()) }
 
     val factRepository by lazy { FactRepository.newInstance() }
