+++ app/src/main/java/com/tmvlg/factorcapgame/data/repository/user/UserRepository.kt
@@ -6,11 +6,13 @@ import com.tmvlg.factorcapgame.data.preferences.PreferenceProvider
 class UserRepository(
     private val preferenceProvider: PreferenceProvider
 ) {
+    //saves statistics to shared pref
     @WorkerThread
     fun saveGame(statistics: Statistics) {
         preferenceProvider.saveStatistics(statistics)
     }
 
+    //loads statistics to shared pref
     @WorkerThread
     fun getStats(): Statistics {
         return preferenceProvider.getStatistics()
