+++ app/src/main/java/com/tmvlg/factorcapgame/ui/multiplayergame/FindLobbyFragment.kt
@@ -30,6 +30,8 @@ class FindLobbyFragment : Fragment() {
         _binding = null
     }
 
+
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         binding.returnButton.setOnClickListener {
