+++ app/src/main/java/com/tmvlg/factorcapgame/ui/multiplayergame/MultiplayerGameViewModelFactory.kt
@@ -3,13 +3,12 @@ package com.tmvlg.factorcapgame.ui.multiplayergame
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.tmvlg.factorcapgame.data.repository.fact.FactRepository
-import com.tmvlg.factorcapgame.data.repository.game.GameRepositoryImpl
+import com.tmvlg.factorcapgame.data.repository.game.GameRepository
 import com.tmvlg.factorcapgame.data.repository.user.UserRepository
-import com.tmvlg.factorcapgame.ui.singlegame.SingleGameViewModel
 import java.lang.RuntimeException
 
 class MultiplayerGameViewModelFactory (
-    private val gameRepository: GameRepositoryImpl,
+    private val gameRepository: GameRepository,
     private val factRepository: FactRepository,
     private val userRepository: UserRepository
 ) : ViewModelProvider.Factory {
