+++ app/src/main/java/com/tmvlg/factorcapgame/ui/singlegame/SingleGameViewModel.kt
@@ -1,6 +1,5 @@
 package com.tmvlg.factorcapgame.ui.singlegame
 
-import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.map
@@ -9,34 +8,41 @@ import android.util.Log
 import com.tmvlg.factorcapgame.data.repository.fact.Fact
 import com.tmvlg.factorcapgame.data.repository.fact.FactRepository
 import com.tmvlg.factorcapgame.data.repository.game.Game
-import com.tmvlg.factorcapgame.data.repository.game.GameRepositoryImpl
+import com.tmvlg.factorcapgame.data.repository.game.GameRepository
 import com.tmvlg.factorcapgame.data.repository.user.UserRepository
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import java.io.IOException
 
 class SingleGameViewModel(
-    private val gameRepository: GameRepositoryImpl,
+    private val gameRepository: GameRepository,
     private val factRepository: FactRepository,
     private val userRepository: UserRepository
 ) : ViewModel() {
+    //exception that throws when can't fetch a fact
     private val _exception = MutableLiveData<IOException?>(null)
     val exception = _exception.map { it }
 
+    //is game already finished
     private val _gameFinished = MutableLiveData(false)
     val gameFinished = _gameFinished.map { it }
 
+    //is your score higher than previous scores
     private val _isHighScore = MutableLiveData(false)
     val isHighScore = _isHighScore.map { it }
 
+    //how much answers were correct
     private val _rightAnswersCount = MutableLiveData(0)
     val rightAnswersCount = _rightAnswersCount.map { it }
 
+    //fetches fact object from repository
     private val _fact = MutableLiveData<Fact>()
     val fact = _fact.map { it }
 
+    //how much game longs. Stores in statistics at game history
     private var timeElapsed: Long = 0
 
+    //check is answer is correct. If true loads new fact. If false finishes a game and saves statistics
     fun sendAnswer(answer: Boolean)= viewModelScope.launch {
         if (fact.value?.isTrue == answer) {
             _rightAnswersCount.postValue(rightAnswersCount.value?.plus(1))
@@ -48,6 +54,7 @@ class SingleGameViewModel(
         }
     }
 
+    //calls to save statistics in Statistics object. Shows in statistics fragment
     fun saveStats() = viewModelScope.launch {
         val stats = userRepository.getStats()
         val score = rightAnswersCount.value ?: 0
@@ -62,6 +69,7 @@ class SingleGameViewModel(
         userRepository.saveGame(stats)
     }
 
+    //fetches new fact from repository
     fun getFact() = viewModelScope.launch {
         try {
             _fact.postValue(factRepository.getFact())
@@ -70,6 +78,8 @@ class SingleGameViewModel(
             _exception.postValue(e)
         }
     }
+
+    //saves current game to DB. Shows in game history
     fun saveGame() = viewModelScope.launch {
         val score = (_rightAnswersCount.value ?: 0).toLong()
         val game = Game(score, timeElapsed)
@@ -77,6 +87,7 @@ class SingleGameViewModel(
         Log.d("1", "saveGame: allGames=${gameRepository.allGames}")
     }
 
+    //fetches first fact and starts a timer to count game duration
     private fun startGame() {
         getFact()
         viewModelScope.launch {
