+++ app/src/main/java/com/tmvlg/factorcapgame/ui/singlegame/SingleGameViewModelFactory.kt
@@ -3,12 +3,13 @@ package com.tmvlg.factorcapgame.ui.singlegame
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.tmvlg.factorcapgame.data.repository.fact.FactRepository
-import com.tmvlg.factorcapgame.data.repository.game.GameRepositoryImpl
+import com.tmvlg.factorcapgame.data.repository.game.GameRepository
 import com.tmvlg.factorcapgame.data.repository.user.UserRepository
 import java.lang.RuntimeException
 
+//this class needed to pass arguments to the viewmodel
 class SingleGameViewModelFactory(
-    private val gameRepository: GameRepositoryImpl,
+    private val gameRepository: GameRepository,
     private val factRepository: FactRepository,
     private val userRepository: UserRepository
 ) : ViewModelProvider.Factory {
