+++ app/src/main/res/layout/fragment_invite.xml
@@ -24,7 +24,7 @@
             android:id="@+id/tv_invite_friend"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="Invite a Friend"
+            android:text="@string/invite_a_friend"
             android:textSize="45sp"
             android:layout_margin="18dp"
             android:fontFamily="@font/norwester"
@@ -97,7 +97,7 @@
             android:textAllCaps="false"
             android:letterSpacing="0"
             android:padding="13dp"
-            android:text="Confirm"
+            android:text="@string/confirm"
             android:textSize="35sp"
             android:textColor="@color/font_color"
             android:fontFamily="@font/norwester"
