+++ app/src/main/res/layout/fragment_no_internet.xml
@@ -1,0 +1,80 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:tools="http://schemas.android.com/tools"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:background="@drawable/ic_wallpaper"
+        android:padding="32dp">
+
+    <TextView
+            android:id="@+id/tv_error_message"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="No Internet connection!"
+            android:textSize="50sp"
+            android:fontFamily="@font/norwester"
+            android:textColor="@color/font_color"
+            android:layout_marginTop="50dp"
+            app:autoSizeTextType="uniform"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            android:textAlignment="center">
+
+    </TextView>
+
+    <ImageView
+            android:id="@+id/error_icon"
+            android:layout_width="125dp"
+            android:layout_height="125dp"
+            android:src="@drawable/ic_no_connection"
+            app:layout_constraintTop_toBottomOf="@+id/tv_error_message"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintBottom_toTopOf="@+id/tv_error_description"
+            app:layout_constraintVertical_chainStyle="packed" />
+
+    <TextView
+            android:id="@+id/tv_error_description"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            android:text="@string/please_check_your_internet_settings_and_return_to_menu"
+            android:textSize="20sp"
+            android:layout_margin="40dp"
+            android:fontFamily="@font/norwester"
+            android:textColor="@color/font_color"
+            app:autoSizeTextType="uniform"
+            android:textAlignment="center"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintBottom_toTopOf="@+id/menu_button"
+            app:layout_constraintTop_toBottomOf="@id/error_icon">
+
+    </TextView>
+
+    <Button
+            android:id="@+id/menu_button"
+            android:layout_width="270dp"
+            android:layout_height="wrap_content"
+            app:cornerRadius="80dp"
+            android:textAllCaps="false"
+            android:letterSpacing="0"
+            android:padding="13dp"
+            android:layout_marginBottom="24dp"
+            android:text="@string/menu"
+            android:textSize="35sp"
+            android:textColor="@color/font_color"
+            android:fontFamily="@font/norwester"
+            android:src="@drawable/green_round_background"
+            android:theme="@style/Button.Main"
+            android:maxLines="1"
+            app:autoSizeTextType="uniform"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintEnd_toEndOf="parent">
+
+    </Button>
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
